refactor(rest): seperate interface classes into own headers
