+++ google/cloud/google_cloud_cpp_rest_internal.bzl
@@ -20,8 +20,11 @@ google_cloud_cpp_rest_internal_hdrs = [
     "internal/binary_data_as_debug_string.h",
     "internal/curl_handle.h",
     "internal/curl_handle_factory.h",
+    "internal/curl_http_payload.h",
     "internal/curl_impl.h",
     "internal/curl_options.h",
+    "internal/curl_rest_client.h",
+    "internal/curl_rest_response.h",
     "internal/curl_wrappers.h",
     "internal/http_payload.h",
     "internal/make_jwt_assertion.h",
@@ -50,9 +53,11 @@ google_cloud_cpp_rest_internal_srcs = [
     "internal/binary_data_as_debug_string.cc",
     "internal/curl_handle.cc",
     "internal/curl_handle_factory.cc",
+    "internal/curl_http_payload.cc",
     "internal/curl_impl.cc",
+    "internal/curl_rest_client.cc",
+    "internal/curl_rest_response.cc",
     "internal/curl_wrappers.cc",
-    "internal/http_payload.cc",
     "internal/make_jwt_assertion.cc",
     "internal/oauth2_access_token_credentials.cc",
     "internal/oauth2_anonymous_credentials.cc",
@@ -67,8 +72,6 @@ google_cloud_cpp_rest_internal_srcs = [
     "internal/oauth2_refreshing_credentials_wrapper.cc",
     "internal/oauth2_service_account_credentials.cc",
     "internal/openssl_util.cc",
-    "internal/rest_client.cc",
     "internal/rest_request.cc",
-    "internal/rest_response.cc",
     "internal/unified_rest_credentials.cc",
 ]
