+++ google/cloud/google_cloud_cpp_rest_internal_emulator_integration_tests.bzl
@@ -17,5 +17,5 @@
 """Automatically generated unit tests list - DO NOT EDIT."""
 
 google_cloud_cpp_rest_internal_emulator_integration_tests = [
-    "internal/rest_client_integration_test.cc",
+    "internal/curl_rest_client_integration_test.cc",
 ]
