+++ google/cloud/google_cloud_cpp_rest_internal_unit_tests.bzl
@@ -20,9 +20,11 @@ google_cloud_cpp_rest_internal_unit_tests = [
     "internal/binary_data_as_debug_string_test.cc",
     "internal/curl_handle_factory_test.cc",
     "internal/curl_handle_test.cc",
+    "internal/curl_http_payload_test.cc",
     "internal/curl_impl_test.cc",
+    "internal/curl_rest_client_test.cc",
+    "internal/curl_rest_response_test.cc",
     "internal/curl_wrappers_test.cc",
-    "internal/http_payload_test.cc",
     "internal/make_jwt_assertion_test.cc",
     "internal/oauth2_access_token_credentials_test.cc",
     "internal/oauth2_anonymous_credentials_test.cc",
@@ -35,8 +37,6 @@ google_cloud_cpp_rest_internal_unit_tests = [
     "internal/oauth2_refreshing_credentials_wrapper_test.cc",
     "internal/oauth2_service_account_credentials_test.cc",
     "internal/openssl_util_test.cc",
-    "internal/rest_client_test.cc",
     "internal/rest_request_test.cc",
-    "internal/rest_response_test.cc",
     "internal/unified_rest_credentials_test.cc",
 ]
