+++ google/cloud/internal/ci/run_integration_tests_emulator_bazel.sh
@@ -52,7 +52,7 @@ emulator_args=(
 # We need to forward some environment variables suitable for running against
 # the emulator.
 "${BAZEL_BIN}" "${BAZEL_VERB}" "${bazel_test_args[@]}" "${emulator_args[@]}" \
-  "//google/cloud:internal_rest_client_integration_test"
+  "//google/cloud:internal_curl_rest_client_integration_test"
 
 exit_status=$?
