+++ google/cloud/internal/curl_rest_client_integration_test.cc
