+++ google/cloud/internal/curl_rest_client_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/internal/rest_client.h"
+#include "google/cloud/internal/curl_rest_client.h"
 #include <gmock/gmock.h>
 
 namespace google {
