+++ google/cloud/internal/rest_response.h
@@ -26,8 +26,6 @@ namespace cloud {
 namespace rest_internal {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 
-class CurlImpl;
-
 enum HttpStatusCode : std::int32_t {
   kMinContinue = 100,
   kMinSuccess = 200,
@@ -78,28 +76,6 @@ class RestResponse {
   // Creates a HttpPayload object from the underlying HTTP response,
   // invalidating the current RestResponse object.
   virtual std::unique_ptr<HttpPayload> ExtractPayload() && = 0;
-};
-
-// Implements RestResponse using libcurl.
-class CurlRestResponse : public RestResponse {
- public:
-  ~CurlRestResponse() override = default;
-
-  CurlRestResponse(CurlRestResponse const&) = delete;
-  CurlRestResponse(CurlRestResponse&&) = default;
-  CurlRestResponse& operator=(CurlRestResponse const&) = delete;
-  CurlRestResponse& operator=(CurlRestResponse&&) = default;
-
-  HttpStatusCode StatusCode() const override;
-  std::multimap<std::string, std::string> Headers() const override;
-  std::unique_ptr<HttpPayload> ExtractPayload() && override;
-
- private:
-  friend class CurlRestClient;
-  CurlRestResponse(Options options, std::unique_ptr<CurlImpl> impl);
-
-  std::unique_ptr<CurlImpl> impl_;
-  Options options_;
 };
 
 /**
